+++ tools.cpp
@@ -21,7 +21,7 @@ std::vector<uint8_t> hexStrToV8(std::string str) {
 	return v;
 }
 
-std::vector<uint64_t> generatePrimeTable(const uint64_t limit) {
+std::vector<uint32_t> generatePrimeTable(const uint32_t limit) {
 	if (limit < 2) return {};
 	std::vector<uint64_t> compositeTable((limit + 127ULL)/128ULL, 0ULL); // Booleans indicating whether an odd number is composite: 0000100100101100...
 	for (uint64_t f(3ULL) ; f*f <= limit ; f += 2ULL) { // Eliminate f and its multiples m for odd f from 3 to square root of the limit
@@ -29,7 +29,7 @@ std::vector<uint64_t> generatePrimeTable(const uint64_t limit) {
 		for (uint64_t m((f*f) >> 1ULL) ; m <= (limit >> 1ULL) ; m += f) // Start eliminating at f^2 (multiples of f below were already eliminated)
 			compositeTable[m >> 6ULL] |= 1ULL << (m & 63ULL);
 	}
-	std::vector<uint64_t> primeTable(1, 2);
+	std::vector<uint32_t> primeTable(1, 2);
 	for (uint64_t i(1ULL) ; (i << 1ULL) + 1ULL <= limit ; i++) { // Fill the prime table using the composite table
 		if (!(compositeTable[i >> 6ULL] & (1ULL << (i & 63ULL))))
 			primeTable.push_back((i << 1ULL) + 1ULL); // Add prime number 2i + 1
